refactor(app): 拆分 AssetLibrary、AssetInspector、Sidebar
提取工具栏、表格、批量栏、上下文菜单及 inspector/sidebar 子组件，降低巨石文件维护成本。
