+++ app/src/features/inspector/index.ts
@@ -1,3 +1,4 @@
 export { AssetInspector } from './AssetInspector';
+export type { MetadataReviewSession } from './inspectorTypes';
 export { default as ImageEditModal } from './ImageEditModal';
 export { default as ImagePreviewModal } from './ImagePreviewModal';
